Fix pr issues

+++ include/vcml/debugging/target.h
@@ -41,9 +41,9 @@ struct cpureg {
     cpureg(u64 no, const string& nm, u64 sz, u64 cnt, int p):
         regno(no), name(nm), size(sz), count(cnt), prot(p), host() {}
 
-    u64 width() const { return size * 8; }
-    u64 total_size() const { return size * count; }
-    u64 total_width() const { return total_size() * 8; }
+    size_t width() const { return size * 8; }
+    size_t total_size() const { return size * count; }
+    size_t total_width() const { return total_size() * 8; }
 
     bool is_readable() const { return is_read_allowed(prot); }
     bool is_writeable() const { return is_write_allowed(prot); }
